Move Volume definition to _space_api.py
